+++ rts/Lua/LuaHandle.h
@@ -171,11 +171,9 @@ class CLuaHandle : public CEventClient
 		void UnitMoveFailed(const CUnit* unit) override;
 		void UnitArrivedAtGoal(const CUnit* unit) override;
 		
-		// ---- Self Destruct Events  ----
 		void UnitSelfDestructStarted(const CUnit* unit) override;
 		void UnitSelfDestructCancelled(const CUnit* unit) override;
-		void UnitSelfDestructProgress(const CUnit* unit, int remainingSeconds) override;
-		// ---- End Self Destruct Events  ----
+		void UnitSelfDestructProgress(const CUnit* unit, float remainingSeconds) override;
 
 		void RenderUnitDestroyed(const CUnit* unit) override;
